requirements.txt
streamlit
python-dotenv
st_clickable_images
streamlit_extras
ibm-watson-machine-learning